be slightly safer (if racy):  check if each argument is a symlink before clobbering it
git-svn-id: https://subversion01.broadinstitute.org/APSG@1251 95e82132-2061-11dd-9cc4-33af95911ecc

+++ scripts/ltouch.sh
@@ -2,5 +2,10 @@
 
 for link in $*
 do
-    ln -sf `readlink $link` $link
+    if [[ -L $link ]]
+    then
+	ln -sf `readlink $link` $link
+    else
+	echo skipping $link:  not a symbolic link >&2
+    fi
 done
